SG-2004-04-20-1: Bugfix:    - fixed the build for DARWIN                             - set a max limit for MAX_DYN_EC in qmaster                             - updated the profiling in the scheduler                 Changed:   sge_schedd, sge_qmaster                 Review:    CR

+++ source/3rdparty/qmake/remote-sge.c
@@ -823,7 +823,7 @@ static void unlock_hostentry(off_t offset) {
       lock_hostfile();
    
       if(be_verbose) {
-         fprintf(stderr, "unlock_hostentry %d\n", offset);
+         fprintf(stderr, "unlock_hostentry %d\n", (int) offset);
       }
    
       if(lseek(hostfile, offset * (MAXHOSTNAMELEN + sizeof(char)) + sizeof(hostfile_info), SEEK_SET) < 0) {
